requirements.txt
langchain
openai
streamlit
python-dotenv
tiktoken
streamlit-chat